turbo-tasks-backend: share persistent-backend test setup, trim parent_count tests
Extracts the hand-rolled backend construction that parent_count.rs and
eviction.rs each carried into `tests/util.rs`. Both need a real on-disk backend
and `tt.backend()` for the `*_for_testing` hooks, which `register!` can't
provide — its harness hands back an `Arc<dyn TurboTasksApi>` that erases the
concrete type, and it owns the session loop these tests need to drive
themselves.

Drops two parent_count tests that were not pulling their weight:

- `parent_count_decrement_survives_snapshot_evict` — subsumed by
  `parent_count_survives_reopen`, which round-trips through a full DB reopen
  rather than just evict + restore.
- `parent_count_wide_fanout_parallel_path` — built a 12,000-task graph to cover
  the chunked `connect_children` path, but the `+1` lives in
  `process_new_children`, which both the serial and parallel paths call with
  disjoint chunks; the parent-side `extend_children` runs once, serially, before
  any chunking. It also asserted on only 5 of 12,000 children, so it did not
  actually detect a seeded chunk-boundary bug that the small tests both caught.
